separated data models from state management models into a new file api_models.dart
